backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
google-auth
google-api-python-client
